scripts/requirements.txt
feedparser>=6.0.11
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
beautifulsoup4>=4.12.2
tzdata>=2024.1

